+++ axum-core/CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- None.
+- **added:** Implement `IntoResponseParts` for `()` ([#2471])
+
+[#2471]: https://github.com/tokio-rs/axum/pull/2471
 
 # 0.4.2 (29. December, 2023)